Update directory structure + add 2nd CNN experiment
- Exp 4: Simpified CNN with Early Stopping
